etl/postgres_to_es/requirements.txt
elastic-transport==8.4.0
elasticsearch==7.9.1
flake8==4.0.0
python-dotenv==0.20.0
psycopg2-binary==2.9.3
pydantic==1.10.2
urllib3==1.26.14